requirements.txt
ipython>=7.14.0
PyVISA>=1.10.1
matplotlib>=3.2.1
engineering-notation>=0.6.0